feat: divide retrieval and logreg
