+++ autointent/modules/embedding/__init__.py
@@ -1,5 +1,6 @@
 """These modules are used only for optimization as they use proxy metrics for choosing best embedding model."""
 
-from ._retrieval import LogRegEmbedding, RetrievalEmbedding
+from ._logreg import LogRegEmbedding
+from ._retrieval import RetrievalEmbedding
 
 __all__ = ["LogRegEmbedding", "RetrievalEmbedding"]
